+++ minicbor-tests/tests/identities.rs
@@ -85,9 +85,9 @@ fn f16() {
     fn property(arg: f32) -> bool {
         let mut e = Encoder::new(Vec::new());
         e.f16(arg).unwrap();
-        let mut d = Decoder::new(e.as_ref());
+        let mut d = Decoder::new(e.writer());
         let val = d.f16().unwrap();
-        assert_eq!(d.position(), e.as_ref().len());
+        assert_eq!(d.position(), e.writer().len());
         half::f16::from_f32(arg).to_f32().to_bits() == val.to_bits()
     }
     quickcheck(property as fn(f32) -> bool)
@@ -277,24 +277,8 @@ fn ip() {
 }
 
 #[test]
-fn legacy_ip() {
-    fn property(x: std::net::IpAddr) -> bool {
-        identity(minicbor::legacy::IpAddr(x))
-    }
-    quickcheck(property as fn(std::net::IpAddr) -> bool)
-}
-
-#[test]
 fn ipv4() {
     quickcheck(identity as fn(std::net::Ipv4Addr) -> bool)
-}
-
-#[test]
-fn legacy_ipv4() {
-    fn property(x: std::net::Ipv4Addr) -> bool {
-        identity(minicbor::legacy::Ipv4Addr(x))
-    }
-    quickcheck(property as fn(std::net::Ipv4Addr) -> bool)
 }
 
 #[test]
@@ -303,24 +287,8 @@ fn ipv6() {
 }
 
 #[test]
-fn legacy_ipv6() {
-    fn property(x: std::net::Ipv6Addr) -> bool {
-        identity(minicbor::legacy::Ipv6Addr(x))
-    }
-    quickcheck(property as fn(std::net::Ipv6Addr) -> bool)
-}
-
-#[test]
 fn socketaddr() {
     quickcheck(identity as fn(std::net::SocketAddr) -> bool)
-}
-
-#[test]
-fn legacy_socketaddr() {
-    fn property(x: std::net::SocketAddr) -> bool {
-        identity(minicbor::legacy::SocketAddr(x))
-    }
-    quickcheck(property as fn(std::net::SocketAddr) -> bool)
 }
 
 #[test]
@@ -329,29 +297,11 @@ fn socketaddrv4() {
 }
 
 #[test]
-fn legacy_socketaddrv4() {
-    fn property(x: std::net::SocketAddrV4) -> bool {
-        identity(minicbor::legacy::SocketAddrV4(x))
-    }
-    quickcheck(property as fn(std::net::SocketAddrV4) -> bool)
-}
-
-#[test]
 fn socketaddrv6() {
     fn property(mut x: std::net::SocketAddrV6) -> bool {
         x.set_flowinfo(0);
         x.set_scope_id(0);
         identity(x)
-    }
-    quickcheck(property as fn(std::net::SocketAddrV6) -> bool)
-}
-
-#[test]
-fn legacy_socketaddrv6() {
-    fn property(mut x: std::net::SocketAddrV6) -> bool {
-        x.set_flowinfo(0);
-        x.set_scope_id(0);
-        identity(minicbor::legacy::SocketAddrV6(x))
     }
     quickcheck(property as fn(std::net::SocketAddrV6) -> bool)
 }